refactor: task の CRUD 処理を App に担わせるため
addTask 関数の定義場所を TaskInput から App に変更
